requirements.txt
asgiref==3.5.0
attrs==21.4.0
Django==4.0.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
pycountry==22.3.5